+++ BrokenTC2/assets/update/changelog.md
@@ -1,5 +1,5 @@
 # New Features
-- It is now possible to switch mode from a keyboard key
+- It is now possible to select on which screen the gear display should be displayed
 
 <br/><br/>